Move JsonSerializable to json module (#2300)
